Separate plugin loading from event management
Plugin loading (and, later, unloading) should be handled separately
from registration of event handlers and dispatch of events. To that
end, what was formerly known as the event manager will now be split
into two modules: event_manager.ex, responsible for event routing,
and plugin_loader.ex, responsible for managing the plugins that are
active on the server at any given moment in time.