If config file is absent, write default config to disk
In the event that '#{ALEMBIC_HOME}/config.exs' does not
already exist, the Config module will now retrieve the
default configuration object, use that as the config,
and write its contents to the standard config file path.

Note that this implementation is relatively naïve; no
attempt is made to correctly format the written file for
readability. This should be addressed in the future.