Defer handing over control of sockets to client processes
Rather than immediately changing the controlling process
of a socket representing a newly connected client to the
client process, allow the client process to take control
of the socket on its own terms.

This allows the client process to delegate control of the
socket to the reader process that it spawns upon first
spinning up without requiring an extra handoff of socket
control in the intermediary.