Initial draft of Minecraft interoperability
The eventual plan is for interoperability with particular
client protocols to be implemented in such a way that almost
any voxel game's client could be plugged into an Alembic
server, provided that server was running an adapter for the
client's protocol.

The Alembic.Minecraft module namespace contains code that will
serve as the reference implementation for such an adapter,
translating server-side events into Minecraft-protocol packets
and vice versa on behalf of those clients who are using the
Minecraft protocol to connect.