Add preliminary typespecs to Translator module
At some point in the future, a separate module will probably be
created to wrap more explicit specifications of the 'request'
and 'event' types in particular. These types will likely end up
being defined in terms of their subtypes, i.e. the exact sorts
of requests and events that make it into the standard API.