Document new protocol-agnostic cleverness
As of now, we're going to determine the protocol being used by a
particular client by associating each client type with a particular
port number in the config. A client that wants to use a particular
protocol will simply connect to the appropriate port; then, the TCP
server can simply fetch the appropriate translator module and run
all future communication with the client in question through said
translator.

This is probably the most widely-accepted way of doing stuff, and
I'm honestly pretty embarassed that it only just now occurred to me.
I'm a networking noob, ok? :P