Refactor Client module for generality
We're finally able to remove all specific references to the
Minecraft protocol translator module from the generic Client
module; the Client module, like many of the other modules we've
recently refactored, is now protocol-agnostic.

I've also taken this opportunity to clean up some of the ExActor
code in the Client module, increasing the readability of said
code by a good amount.