Slightly different approach to config evaluation
The 'config.exs' file will now be evaluated as a series
of bindings, rather than a single keyword list. The end
result is the same (the bindings are translated into a
keyword list automatically by 'Code.eval_string/3'), but
it should now be possible to treat the config file as an
Elixir script if dynamic configuration is desired.