Write config file in new format, not old
When writing the default configuration options to a file,
the new format (a number of variable bindings) rather than
the old format (a single keyword list) should be used.

I've also used this as an opportunity to prettify the output
of writing the default config object to a file; newlines are
now inserted where they belong and I could potentially swap
out 'Kernel.inspect/2' for my own function down the road if
I want to change the way a certain type of value is treated
during stringification.