Remove dependency on ex_doc
We're not really using ex_doc for documentation at the
moment, so its presence in the Mixfile is kind of
unnecessary. We might switch back to using it at a later
date, but I don't like the way it plays with macros that
generate functions so far – especially since I use ExActor
so heavily throughout the codebase.